+++ History.md
@@ -1,7 +1,17 @@
+2.0.3
+=======
+
+
+- #201 correct embedded files in XRechnung
+
 2.0.2
 =======
+2020-11-25
+
 - #197 file attachments
 - Charges/Allowances CategoryCode improvement PR #198 Thanks to weclapp-dev
+- support reverse charge
+- support intra community supply also in product class, not only interface
 
 2.0.1
 =====